requirements.txt
asgiref==3.8.1
Django==5.1.3
django-environ==0.11.2
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.2
whitenoise==6.8.2
